+++ index.js
@@ -49,6 +49,14 @@ Sifaka.prototype.debug = function (key, message) {
         this.debugLogger(new Date() * 1 + "\t" + message);
     }
 };
+Sifaka.prototype.exists = function (key, options, callback) {
+    if(arguments.length === 2 && typeof arguments[1] === "function") {
+        callback = options;
+        options = {};
+    }
+    this.backend.exists(key, options, callback);
+};
+
 Sifaka.prototype.get = function (key, workFn, options, callback) {
     var self = this;
 
@@ -56,13 +64,16 @@ Sifaka.prototype.get = function (key, workFn, options, callback) {
         callback = options;
         options = {};
     }
+    if(options.metaOnly && ["hit", "miss"].indexOf(options.metaOnly) == -1) {
+        throw new RangeError("options.metaOnly should be one of hit or miss");
+    }
 
     self.debug(key, "GET");
 
     if(this.locks[key]) {
         // This node is already polling the lock on this key
         self.debug(key, "CACHE MISS - LOCK PENDING");
-        self._addPending(key, callback);
+        self._addPending(key, callback, options);
     } else {
 
         self.backend.get(key, options, function (err, data, state) {
@@ -77,26 +88,42 @@ Sifaka.prototype.get = function (key, workFn, options, callback) {
                 self.stats.hit++;
                 self.debug(key, "CACHE HIT");
 
-                self._deserialize(data, function (deserializeErr, serializedData) {
-                    callback(err || deserializeErr, serializedData);
-                });
+                if(options.metaOnly === "hit") {
+                    // Pass "data" through here - so that we can verify in the tests that the backends are returning
+                    // undefined (and hopefully not fetching data from the store)
+                    callback(err, data, state);
+                } else {
+                    self._deserialize(data, function (deserializeErr, serializedData) {
+                        callback(err || deserializeErr, serializedData, state);
+                    });
+                }
 
                 // check if we need to refresh the data in the background
                 if(state.stale) {
                     self.stats.stale++;
                     self.backend.lock(key, null, function (err, acquired) {
                         if(acquired) {
-                            self._doWork(key, options, workFn);
+                            self.debug(key, "GOT LOCK FOR STALE REFRESH");
+
+                            self._doWork(key, options, workFn, state);
                         }
+
                     });
                 }
             } else {
                 self.stats.miss++;
-                self._addPending(key, callback);
+                if(options.metaOnly === "miss") {
+                    // We don't need to wait for a result to be calculated, or trigger one
+                    self.debug(key, "META ONLY CACHE MISS");
+                    return callback(err, void 0, state);
+                } else {
+                    self._addPending(key, callback, options);
+                }
                 self.debug(key, "CACHE MISS");
                 self.backend.lock(key, null, function (err, acquired) {
                     if(acquired === true) {
-                        self._doWork(key, options, workFn);
+                        self.debug(key, "GOT LOCK");
+                        self._doWork(key, options, workFn, state);
                     } else {
                         self.debug(key, "WAITING FOR LOCK");
                         self.locks[key] = self._watchLock(key); // We're already aware of this lock being held
@@ -106,11 +133,11 @@ Sifaka.prototype.get = function (key, workFn, options, callback) {
         });
     }
 };
-Sifaka.prototype._addPending = function (key, callback) {
+Sifaka.prototype._addPending = function (key, callback, options) {
     this.debug(key, "PENDING ADDED");
 
     this.pendingCallbacks[key] = this.pendingCallbacks[key] || [];
-    this.pendingCallbacks[key].push(callback);
+    this.pendingCallbacks[key].push({options: options, cb: callback});
 };
 Sifaka.prototype._calculateCacheTimes = function (key, duration, data, options, callback) {
     options = options || {};
@@ -131,7 +158,7 @@ Sifaka.prototype._checkForResult = function (key) {
             self.locks[key] = null;
             self.lockCheckCounts[key] = 0;
             self._deserialize(data, function (err, serializedData) {
-                self._resolvePendingCallbacks(key, err, serializedData);
+                self._resolvePendingCallbacks(key, err, serializedData, false, state);
             });
         } else {
             self.lockCheckCounts[key] += 1;
@@ -162,27 +189,40 @@ Sifaka.prototype._deserialize = function (data, callback) {
     }
 };
 
-Sifaka.prototype._doWork = function (key, options, workFunction, callback) {
+Sifaka.prototype._doWork = function (key, options, workFunction, state, callback) {
     var self = this;
     self.debug(key, "TRIGGERING WORK");
     var start = new Date();
-    workFunction(function (workError, data) {
+    workFunction(function (workError, data, storedCallback) {
         self.stats.work++;
         var duration = new Date() - start;
         if(workError && !workError.cache) {
+            self.debug(key, "ERROR, NOT CACHED: "+workError+" - UNLOCKING");
+
             self.backend.unlock(key, {}, function () {
-                self._resolvePendingCallbacks(key, workError, data);
+                self.debug(key, "UNLOCKED AFTER WORK THAT ERRORED");
+                self._resolvePendingCallbacks(key, workError, data, true, state);
             });
         } else {
             self._calculateCacheTimes(key, duration, data, options, function (err, cachePolicyResult) {
                 if(!cachePolicyResult.noCache) {
                     self._serialize(data, function (err, serializedData) {
+                        self.debug(key, "STORING AND UNLOCKING....");
                         self.backend.store(key, serializedData, workError, {
                             unlock: true, cachePolicy: cachePolicyResult
-                        }, noop);
+                        }, function (storedError, storedResult) {
+                            self._resolvePendingCallbacks(key, workError, data, true, state);
+                            if(storedCallback) {
+                                storedCallback(storedError, storedResult);
+                            }
+                        });
+                    });
+                } else {
+                    self.debug(key, "UNLOCKING");
+                    self.backend.unlock(key, options, function () {
+                        self._resolvePendingCallbacks(key, workError, data, true, state);
                     });
                 }
-                self._resolvePendingCallbacks(key, workError, data);
 
                 // Needed?
                 if(callback) {
@@ -204,19 +244,31 @@ Sifaka.prototype._logStats = function () {
     this.stats = {hit: 0, miss: 0, work: 0};
     this.emit("stats", currentStats);
 };
-Sifaka.prototype._resolvePendingCallbacks = function (key, err, data) {
+Sifaka.prototype._resolvePendingCallbacks = function (key, err, data, didWork, state) {
     var self = this;
     if(self.locks[key]) {
         clearTimeout(self.locks[key]);
         self.locks[key] = null;
-        self.backend.unlock(key, {}, noop);
+        if(!didWork) {
+            self.debug(key, "UNLOCKING");
+
+            self.backend.unlock(key, {}, noop);
+        }
     }
 
     this.debug(key, "RESOLVING PENDING CALLBACKS");
     this.pendingCallbacks[key] = this.pendingCallbacks[key] || [];
+    state.pending = true;
     while(this.pendingCallbacks[key].length) {
-        var cb = this.pendingCallbacks[key].shift();
-        cb(err, data);
+        var pending = this.pendingCallbacks[key].shift();
+        var cb = pending.cb;
+        var options = pending.options;
+
+        if(options.metaOnly && options.metaOnly == "miss") {
+            cb(err, void 0, state);
+        } else {
+            cb(err, data, state);
+        }
     }
 };
 Sifaka.prototype._watchLock = function (key) {